Breakout layouts into their own folder and scripts and add option to hide layout from command palette for more testing-oriented layouts
