+++ src/vs/workbench/services/positronLayout/browser/layouts/positronFourPaneDsLayout.ts
@@ -1,0 +1,60 @@
+/*---------------------------------------------------------------------------------------------
+ *  Copyright (C) 2024 Posit Software, PBC. All rights reserved.
+ *--------------------------------------------------------------------------------------------*/
+
+import { localize2 } from 'vs/nls';
+import { registerAction2 } from 'vs/platform/actions/common/actions';
+import { Parts } from 'vs/workbench/services/layout/browser/layoutService';
+import { PositronLayoutAction, PositronLayoutInfo } from './layoutAction';
+
+
+export const positronFourPaneDsLayout: PositronLayoutInfo = {
+	id: 'workbench.action.positronFourPaneDataScienceLayout',
+	codicon: 'positron-four-pane-ds-layout',
+	label: localize2('choseLayout.stacked', 'Stacked Layout'),
+	layoutDescriptor: {
+		[Parts.SIDEBAR_PART]: {
+			size: '15%',
+			hidden: false,
+		},
+		[Parts.PANEL_PART]: {
+			size: '40%',
+			hidden: false,
+			alignment: 'center',
+			viewContainers: [
+				{
+					id: 'workbench.panel.positronConsole',
+					opened: true,
+				},
+				{
+					id: 'terminal',
+				},
+			]
+		},
+		[Parts.AUXILIARYBAR_PART]: {
+			size: '30%',
+			hidden: false,
+			viewContainers: [
+				{
+					id: 'workbench.panel.positronSession',
+					opened: true,
+					views: [
+						{
+							id: 'workbench.panel.positronVariables',
+						},
+						{
+							id: 'workbench.panel.positronPlots',
+						},
+					]
+				},
+			]
+		}
+	},
+};
+
+
+registerAction2(class extends PositronLayoutAction {
+	constructor() {
+		super(positronFourPaneDsLayout);
+	}
+});
