+++ src/vs/workbench/services/positronLayout/browser/positronCustomViews.ts
@@ -1,245 +1,19 @@
 /*---------------------------------------------------------------------------------------------
  *  Copyright (C) 2024 Posit Software, PBC. All rights reserved.
  *--------------------------------------------------------------------------------------------*/
-
-
-import { ILocalizedString, localize2 } from 'vs/nls';
-import { Categories } from 'vs/platform/action/common/actionCommonCategories';
-import { Action2, registerAction2 } from 'vs/platform/actions/common/actions';
-import { ServicesAccessor } from 'vs/platform/instantiation/common/instantiation';
 import { IQuickPickItem } from 'vs/platform/quickinput/common/quickInput';
-import { Parts } from 'vs/workbench/services/layout/browser/layoutService';
-import { IPositronLayoutService } from 'vs/workbench/services/positronLayout/browser/interfaces/positronLayoutService';
+import { CustomPositronLayoutDescription } from 'vs/workbench/services/positronLayout/common/positronCustomViews';
+import { positronFourPaneDsLayout } from './layouts/positronFourPaneDsLayout';
+import { positronTwoPaneLayout } from './layouts/positronTwoPaneLayout';
+import { positronNotebookLayout } from './layouts/positronNotebookLayout';
+import { PositronLayoutInfo } from './layouts/layoutAction';
 
-// Need this import or else the PositronLayoutService will not be registered.
+// Imports needed to register the layout service and non-primary layouts. (Otherwise the scripts
+// are not run and the commands are not registered.)
 import 'vs/workbench/services/positronLayout/browser/positronLayoutService';
-import { CustomPositronLayoutDescription, KnownPositronLayoutParts } from 'vs/workbench/services/positronLayout/common/positronCustomViews';
+import './layouts/maximizedPartLayouts';
+import './layouts/positronHelpPaneDocked';
 
-
-type PositronLayoutInfo = {
-	id: string;
-	codicon?: string;
-	label: ILocalizedString;
-	layoutDescriptor: CustomPositronLayoutDescription;
-};
-
-export const positronFourPaneDsLayout: PositronLayoutInfo = {
-	id: 'workbench.action.positronFourPaneDataScienceLayout',
-	codicon: 'positron-four-pane-ds-layout',
-	label: localize2('choseLayout.stacked', 'Stacked Layout'),
-	layoutDescriptor: {
-		[Parts.SIDEBAR_PART]: {
-			size: '15%',
-			hidden: false,
-		},
-		[Parts.PANEL_PART]: {
-			size: '40%',
-			hidden: false,
-			alignment: 'center',
-			viewContainers: [
-				{
-					id: 'workbench.panel.positronConsole',
-					opened: true,
-				},
-				{
-					id: 'terminal',
-				},
-			]
-		},
-		[Parts.AUXILIARYBAR_PART]: {
-			size: '30%',
-			hidden: false,
-			viewContainers: [
-				{
-					id: 'workbench.panel.positronSession',
-					opened: true,
-					views: [
-						{
-							id: 'workbench.panel.positronVariables',
-						},
-						{
-							id: 'workbench.panel.positronPlots',
-						},
-					]
-				},
-			]
-		}
-	},
-};
-
-export const positronTwoPaneLayout: PositronLayoutInfo = {
-	id: 'workbench.action.positronTwoPaneDataScienceLayout',
-	codicon: 'positron-two-pane-ds-layout',
-	label: localize2('choseLayout.sideBySide', 'Side-By-Side Layout'),
-	layoutDescriptor: {
-		[Parts.PANEL_PART]: {
-			hidden: true,
-			alignment: 'center'
-		},
-		[Parts.SIDEBAR_PART]: {
-			hidden: true
-		},
-		[Parts.AUXILIARYBAR_PART]: {
-			hidden: false,
-			size: '40%',
-			viewContainers: [
-				{
-					id: 'workbench.panel.positronSession',
-					opened: true,
-					views: [
-						{
-							id: 'workbench.panel.positronConsole',
-						},
-						{
-							id: 'workbench.panel.positronVariables',
-							collapsed: true,
-						},
-						{
-							id: 'workbench.panel.positronPlots',
-							collapsed: true,
-						},
-					]
-				},
-				{
-					id: 'terminal',
-				}
-			]
-		},
-	},
-};
-
-export const positronNotebookLayout: PositronLayoutInfo = {
-	id: 'workbench.action.positronNotebookLayout',
-	codicon: 'positron-notebook-layout',
-	label: localize2('chooseLayout.notebookLayout', 'Notebook Layout'),
-	layoutDescriptor: {
-		[Parts.PANEL_PART]: {
-			size: '40%',
-			hidden: true,
-			alignment: 'center',
-			viewContainers: [
-				{
-					id: 'workbench.panel.positronConsole',
-					opened: true,
-				},
-				{
-					id: 'terminal',
-				},
-			]
-		},
-		[Parts.SIDEBAR_PART]: {
-			hidden: false,
-			size: '15%',
-			minSize: 180,
-			hideIfBelowMinSize: true,
-		},
-		[Parts.AUXILIARYBAR_PART]: {
-			hidden: true,
-			size: '20%',
-			minSize: 180,
-		},
-	},
-};
-
-
-abstract class PositronLayoutAction extends Action2 {
-	private _layout: CustomPositronLayoutDescription;
-	constructor(
-		layoutInfo: PositronLayoutInfo
-	) {
-		super({
-			id: layoutInfo.id,
-			title: layoutInfo.label,
-			category: Categories.View,
-			f1: true,
-		});
-
-		this._layout = layoutInfo.layoutDescriptor;
-	}
-	run(accessor: ServicesAccessor): void {
-		const positronLayoutService = accessor.get(IPositronLayoutService);
-		positronLayoutService.setLayout(this._layout);
-	}
-}
-registerAction2(class extends PositronLayoutAction {
-	constructor() {
-		super(positronNotebookLayout);
-	}
-});
-
-registerAction2(class extends PositronLayoutAction {
-	constructor() {
-		super(positronTwoPaneLayout);
-	}
-});
-
-registerAction2(class extends PositronLayoutAction {
-	constructor() {
-		super(positronFourPaneDsLayout);
-	}
-});
-
-
-// Layouts that maximize a single part as much as possible.
-function makeMaximizedPartLayout(part: KnownPositronLayoutParts): CustomPositronLayoutDescription {
-	return {
-		[Parts.PANEL_PART]: { hidden: true },
-		[Parts.SIDEBAR_PART]: { hidden: true },
-		[Parts.AUXILIARYBAR_PART]: { hidden: true },
-		[part]: { hidden: false, size: '100%' },
-	};
-}
-
-registerAction2(class extends PositronLayoutAction {
-	constructor() {
-		super({
-			id: 'workbench.action.fullSizedSidebar',
-			label: localize2('chooseLayout.fullSizedSidebarLayout', 'Maximizedd Sidebar Layout'),
-			layoutDescriptor: makeMaximizedPartLayout(Parts.SIDEBAR_PART),
-		});
-	}
-});
-
-registerAction2(class extends PositronLayoutAction {
-	constructor() {
-		super({
-			id: 'workbench.action.fullSizedPanel',
-			label: localize2('chooseLayout.fullSizedPanelLayout', 'Maximized Panel Layout'),
-			layoutDescriptor: makeMaximizedPartLayout(Parts.PANEL_PART),
-		});
-	}
-});
-
-registerAction2(class extends PositronLayoutAction {
-	constructor() {
-		super({
-			id: 'workbench.action.fullSizedAuxiliaryBar',
-			label: localize2('chooseLayout.fullSizedAuxiliaryBarLayout', 'Maximized Auxiliary Bar Layout'),
-			layoutDescriptor: makeMaximizedPartLayout(Parts.AUXILIARYBAR_PART),
-		});
-	}
-});
-
-
-
-// Action to dump json of the current layout to the console for creation of a custom layout.
-// registerAction2(class DumpViewCustomizations extends Action2 {
-
-// 	constructor() {
-// 		super({
-// 			id: 'workbench.action.dumpViewCustomizations',
-// 			title: localize2('dumpViewCustomizations', "Dump view customizations to console"),
-// 			category: Categories.View,
-// 			f1: true,
-// 		});
-// 	}
-
-// 	run(accessor: ServicesAccessor): void {
-// 		console.log(
-// 			JSON.stringify(createPositronCustomLayoutDescriptor(accessor), null, 2)
-// 		);
-// 	}
-// });
 
 type LayoutPick = IQuickPickItem & { layoutDescriptor: CustomPositronLayoutDescription };
